fix: explicitly define server ids
Prior to this, server ids were set purely based on the position in the
inventory. So, if a mistake were made and a server moved up or down in
position then the associated ID would be wrong, and the wrong data would
be used against it in SquadJS's database. Furthermore, if we were to
ever change the format of how servers are in the inventory this would
cause a headache as currently server ids are calculated based on
position in the inventory. Explicitly defining them while, yes, adding
another hard coded value, avoids this headache entirely.